app/requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pipdeptree==2.3.3
pydantic==1.10.2
pymongo==4.3.3
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
